+++ device/Src/buzzer.c
@@ -2,7 +2,7 @@
 #include "tim.h"
 #include "bsp_dwt.h"
 
-void buzzer_beep_ms(uint8_t ms)
+void buzzer_beep_ms(uint16_t ms)
 {
     HAL_TIM_PWM_Start(&htim12, TIM_CHANNEL_2);
     __HAL_TIM_SET_COMPARE(&htim12, TIM_CHANNEL_2, 2000);
